Refactoring SearchMovie component
it renders, and gets the movies

Also keeping track of movies in the App component
